refactor: extract renderers and GitHub helpers from bootstrap-project.ts
Split bootstrap-project.ts into three focused modules:
- bootstrap-renderers.ts: all pure string rendering functions
- bootstrap-github.ts: all GitHub REST API call helpers
- bootstrap-project.ts: orchestration only, re-exports renderer helpers for backward compat
